Delete and edit left

+++ src/StudentCourse.java
@@ -3,17 +3,18 @@ import java.util.List;
 
 public class StudentCourse extends Course{
     int courseScore;
-    float courseGPA;
+    Double courseGPA;
     String courseSession;
     courseStatus courseStatus;
 
-    public StudentCourse(int courseId, String courseName, Track courseTrack, int courseCredit, int courseScore, float courseGPA, String courseSession, courseStatus courseStatus) {
+    public StudentCourse(int courseId, String courseName, Track courseTrack, int courseCredit, int courseScore, Double courseGPA, String courseSession, courseStatus courseStatus) {
         super(courseId, courseName, courseTrack, courseCredit);
         this.courseScore = courseScore;
         this.courseGPA = courseGPA;
         this.courseSession = courseSession;
         this.courseStatus = courseStatus;
     }
+
 
     public StudentCourse() {
     }
@@ -34,11 +35,11 @@ public class StudentCourse extends Course{
         this.courseScore = courseScore;
     }
 
-    public float getCourseGPA() {
+    public Double getCourseGPA() {
         return courseGPA;
     }
 
-    public void setCourseGPA(float courseGPA) {
+    public void setCourseGPA(Double courseGPA) {
         this.courseGPA = courseGPA;
     }
 
@@ -50,7 +51,7 @@ public class StudentCourse extends Course{
         this.courseSession = courseSession;
     }
 
-    public void setAll(int courseScore, int courseGPA, String courseSession, courseStatus courseStatus) {
+    public void setAll(int courseScore, Double courseGPA, String courseSession, courseStatus courseStatus) {
         this.courseScore = courseScore;
         this.courseGPA = courseGPA;
         this.courseSession = courseSession;
